Move desktop_sessions
