+++ src/ansys/aedt/core/internal/desktop_sessions.py
